backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
google-genai==1.5.0
pydantic==2.10.4
websockets==14.1
